requirements.txt
numpy
pyocr
pillow
tesseract
pytesseract